Редизайн APP

+++ server/controllers/fileController.js
@@ -1,4 +1,7 @@
+// Контроллер
 const File = require('../models/File');
+const path = require('path');
+const { deleteFile } = require('../utils/fileUtils');
 
 // Контроллер для добавления файла
 const addFile = async (userId, fileName, filePath) => {
@@ -27,11 +30,25 @@ const getAllFiles = async () => {
 // Контроллер для удаления файла по ID
 const deleteFileById = async (fileId) => {
     return new Promise((resolve, reject) => {
-        File.deleteFileById(fileId, (err, results) => {
+        File.getFileById(fileId, (err, file) => {
             if (err) {
                 return reject(err);
             }
-            resolve(results);
+            if (!file) {
+                return reject(new Error('Файл не найден'));
+            }
+
+            const fileName = path.basename(file.file_path);
+
+            deleteFile(fileName);
+
+            // Удаляем файл из базы данных
+            File.deleteFileById(fileId, (err, results) => {
+                if (err) {
+                    return reject(err);
+                }
+                resolve(results);
+            });
         });
     });
 };
@@ -45,16 +62,20 @@ const uploadFile = async (req, res) => {
     const { originalname, path: filePath } = req.file;
     const userId = req.body.user_id || null;
 
+    const directoryPath = path.dirname(filePath);
+    const folderName = path.basename(filePath);
+
+
     try {
-        const result = await addFile(userId, originalname, filePath);
+        const result = await addFile(userId, originalname, path.basename(directoryPath)+'/'+folderName);
         res.status(200).json({
             message: 'Файл успешно загружен',
             fileId: result.insertId,
-            filePath
+            filePath: path.basename(directoryPath)+'/'+folderName
         });
     } catch (err) {
         console.error('Ошибка при сохранении файла в базу:', err);
-        res.status(500).json({ message: 'Ошибка при сохранении файла' });
+        res.status(500).json({ message: 'Ошибка при сохранении файла в базу данных', error: err });
     }
 };
